+++ app/src/main/res/layout/activity_ma_liste_de_reservations.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MaListeDeReservations">
+    tools:context="org.doranco.parcours.chauffeurs.MaListeDeReservations">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
